moved default css files into styles dir
