+++ src/App.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import logo from './logo.svg';
-import './App.css';
+import './styles/App.css';
+import RegisterForm from './pages/RegsiterForm';
 
 function App() {
   return (
@@ -18,6 +19,7 @@ function App() {
         >
           Learn React
         </a>
+        <RegisterForm />
       </header>
     </div>
   );
